@@ -31,35 +31,40 @@
     "@storybook/react": "^8.4.7",
     "@storybook/react-vite": "^8.4.7",
     "@storybook/test": "^8.4.7",
-    "@testing-library/dom": "^10.4.0",
-    "@testing-library/react": "^16.0.1",
-    "@types/node": "^22.5.5",
-    "@types/react": "^18.3.6",
-    "@types/react-dom": "^18.3.0",
-    "@vitejs/plugin-react": "^4.3.1",
+    "@tsconfig/vite-react": "^3.4.0",
+    "@types/node": "^22.10.5",
+    "@types/react": "^19.0.3",
+    "@types/react-dom": "^19.0.2",
+    "@vitejs/plugin-react": "^4.3.4",
     "autoprefixer": "^10.4.20",
-    "jsdom": "^25.0.0",
-    "postcss": "^8.4.47",
+    "postcss": "^8.4.49",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "rollup-preserve-directives": "^1.1.3",
     "storybook": "^8.4.7",
-    "tailwindcss": "^3.4.14",
+    "tailwindcss": "^3.4.17",
     "typescript": "^5.7.2",
-    "vite": "^5.4.5",
-    "vite-plugin-dts": "^4.2.4",
-    "vitest": "^2.1.1"
+    "vite": "^6.0.7",
+    "vite-plugin-dts": "^4.4.0"
   },
   "peerDependencies": {
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
   },
   "dependencies": {
     "@arcgis/core": "^4.31.6",
-    "@fontsource-variable/public-sans": "^5.1.1",
-    "@fontsource-variable/roboto-mono": "^5.1.0",
+    "@fontsource-variable/public-sans": "^5.1.2",
+    "@fontsource-variable/roboto-mono": "^5.1.1",
     "@fontsource/source-sans-pro": "^5.1.0",
     "@headlessui/react": "^2.2.0",
     "@matsugov/app-config": "workspace:^",
     "clsx": "^2.1.1",
     "resize-observer-polyfill": "^1.5.1"
+  },
+  "postcss": {
+    "plugins": {
+      "tailwindcss": {},
+      "autoprefixer": {}
+    }
   }
 }
