@@ -10,28 +10,33 @@
   },
   "dependencies": {
     "@matsugov/ui": "workspace:^",
-    "next": "^14.2.14",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "next": "15.1.3",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
     "react-loading-skeleton": "^3.5.0",
-    "react-markdown": "^9.0.1",
+    "react-markdown": "^9.0.3",
     "remark-directive": "^3.0.0",
     "remark-directive-rehype": "^0.4.2",
     "remark-gfm": "^4.0.0",
-    "swiper": "^11.1.14"
+    "swiper": "^11.2.0"
   },
   "devDependencies": {
     "@matsugov/tw-config": "workspace:^",
     "@testing-library/dom": "^10.4.0",
-    "@testing-library/react": "^16.0.1",
-    "@types/node": "^22.5.5",
-    "@types/react": "^18.3.6",
-    "@types/react-dom": "^18.3.0",
-    "@vitejs/plugin-react": "^4.3.1",
+    "@testing-library/react": "^16.1.0",
+    "@tsconfig/next": "^2.0.3",
+    "@types/node": "^22.10.5",
+    "@types/react": "19.0.3",
+    "@types/react-dom": "19.0.2",
+    "@vitejs/plugin-react": "^4.3.4",
     "autoprefixer": "^10.4.20",
-    "jsdom": "^25.0.0",
-    "postcss": "^8.4.47",
-    "tailwindcss": "^3.4.14",
-    "vitest": "^2.1.1"
+    "postcss": "^8.4.49",
+    "tailwindcss": "^3.4.17",
+    "vitest": "^2.1.8"
+  },
+  "resolutions": {
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "next": "15.1.3"
   }
 }
