@@ -510,7 +510,7 @@ const uswdsColors = {
 };
 
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: [],
   important: true,
   theme: {
